Create README.md
A Task management software.
A to do list app helps you organize tasks and stay on top of your deadlines.
Tasks are typically put in order by priority or importance in which each task can be checked out and/or deleted in respect to preference. 
You can use it to manage everything from grocery lists to work-related tasks and lot more.